refactor(a1): extract _search_sql_only to core/search.py (BR-CODE-1)
Moves the STOREFRONT_SQL_ONLY search path out of app.py into a new core/search.py:
search_sql_only (the entry) + its search-exclusive _escape_ilike helper and
_ILIKE_WILDCARDS constant. Self-contained: takes a live db client, depends
outward only on core.helpers (is_speculative_event_name, or_ilike_clause). app.py
imports search_sql_only aliased to the historical _search_sql_only; the 3 call
sites are unchanged.

Adds tests/test_core_search.py (5 unit tests, fake db): _escape_ilike wildcard
escaping, blank-query short-circuit (no DB hit), owned-inventory sort bias, and
speculative/inactive event dropping.

app.py 6965 -> 6812 LOC. Full suite green aside from the known env-only deps
(supabase.Client, pyo3 crypto) CI provides.
